refactor: improve setup of type declaration files
