+++ packages/docusaurus-mdx-loader/deps.d.ts
@@ -5,20 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-declare module '@docusaurus/mdx-loader' {
-  import type {Plugin} from 'unified';
-
-  export type RemarkOrRehypePlugin =
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    [Plugin<any[]>, Record<string, unknown>] | Plugin<any[]>;
-  export type RemarkAndRehypePluginOptions = {
-    remarkPlugins: RemarkOrRehypePlugin[];
-    rehypePlugins: RemarkOrRehypePlugin[];
-    beforeDefaultRemarkPlugins: RemarkOrRehypePlugin[];
-    beforeDefaultRehypePlugins: RemarkOrRehypePlugin[];
-  };
-}
-
 // TODO Types provided by MDX 2.0 https://github.com/mdx-js/mdx/blob/main/packages/mdx/types/index.d.ts
 declare module '@mdx-js/mdx' {
   import type {Processor} from 'unified';
