+++ packages/docusaurus-mdx-loader/package.json
@@ -3,7 +3,7 @@
   "version": "2.0.0-beta.9",
   "description": "Docusaurus Loader for MDX",
   "main": "lib/index.js",
-  "types": "src/types.d.ts",
+  "types": "src/mdx-loader.d.ts",
   "publishConfig": {
     "access": "public"
   },
