+++ packages/docusaurus-mdx-loader/tsconfig.json
@@ -5,6 +5,5 @@
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
     "rootDir": "src",
     "outDir": "lib"
-  },
-  "include": ["src/"]
+  }
 }
