+++ packages/docusaurus-plugin-content-docs/deps.d.ts
