+++ packages/docusaurus-plugin-content-docs/src/props.ts
@@ -17,7 +17,7 @@ import type {
   PropSidebarItem,
   PropTagDocList,
   PropTagDocListDoc,
-} from '@docusaurus/plugin-content-docs-types';
+} from '@docusaurus/plugin-content-docs';
 import {compact, keyBy, mapValues} from 'lodash';
 
 export function toSidebarsProp(loadedVersion: LoadedVersion): PropSidebars {
