+++ packages/docusaurus-plugin-debug/src/plugin-debug.d.ts
