+++ packages/docusaurus-plugin-google-analytics/src/module-aliases.d.ts
@@ -6,7 +6,3 @@
  */
 
 /// <reference types="@docusaurus/module-type-aliases" />
-
-interface Window {
-  ga: (command: string, ...fields: string[]) => void;
-}
