+++ packages/docusaurus-plugin-google-gtag/src/gtag.ts
@@ -9,6 +9,22 @@ import ExecutionEnvironment from '@docusaurus/ExecutionEnvironment';
 import siteConfig from '@generated/docusaurus.config';
 import type {ThemeConfig} from '@docusaurus/plugin-google-gtag';
 
+declare global {
+  interface Window {
+    /* eslint-disable camelcase */
+    gtag: (
+      command: string,
+      fields: string,
+      params: {
+        page_title?: string;
+        page_location?: string;
+        page_path?: string;
+      },
+    ) => void;
+    /* eslint-enable camelcase */
+  }
+}
+
 export default (function () {
   if (!ExecutionEnvironment.canUseDOM) {
     return null;
