+++ packages/docusaurus-plugin-google-gtag/src/module-aliases.d.ts
@@ -5,17 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-/* eslint-disable camelcase */
 /// <reference types="@docusaurus/module-type-aliases" />
-
-interface Window {
-  gtag: (
-    command: string,
-    fields: string,
-    params: {
-      page_title?: string;
-      page_location?: string;
-      page_path?: string;
-    },
-  ) => void;
-}
