+++ packages/docusaurus-plugin-ideal-image/src/plugin-ideal-image.d.ts
@@ -46,5 +46,3 @@ declare module '@theme/IdealImage' {
   }
   export default function IdealImage(props: Props): JSX.Element;
 }
-
-declare module '@endiliey/react-ideal-image';
