+++ packages/docusaurus-theme-classic/package.json
@@ -3,7 +3,7 @@
   "version": "2.0.0-beta.9",
   "description": "Classic theme for Docusaurus",
   "main": "lib/index.js",
-  "types": "src/types.d.ts",
+  "types": "src/theme-classic.d.ts",
   "publishConfig": {
     "access": "public"
   },
@@ -30,7 +30,6 @@
     "@docusaurus/theme-common": "2.0.0-beta.9",
     "@docusaurus/types": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
-    "@docusaurus/utils-common": "2.0.0-beta.9",
     "@docusaurus/utils-validation": "2.0.0-beta.9",
     "@mdx-js/mdx": "^1.6.21",
     "@mdx-js/react": "^1.6.21",
