+++ packages/docusaurus-theme-classic/src/module-aliases.d.ts
