+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/index.tsx
@@ -22,7 +22,7 @@ import type {Props, DocSidebarItemsProps} from '@theme/DocSidebarItem';
 import type {
   PropSidebarItemCategory,
   PropSidebarItemLink,
-} from '@docusaurus/plugin-content-docs-types';
+} from '@docusaurus/plugin-content-docs';
 
 import styles from './styles.module.css';
