+++ packages/docusaurus-theme-classic/src/theme/DocTagDocListPage/index.tsx
@@ -10,7 +10,7 @@ import React from 'react';
 import Layout from '@theme/Layout';
 import Link from '@docusaurus/Link';
 import {ThemeClassNames, usePluralForm} from '@docusaurus/theme-common';
-import type {PropTagDocListDoc} from '@docusaurus/plugin-content-docs-types';
+import type {PropTagDocListDoc} from '@docusaurus/plugin-content-docs';
 import Translate, {translate} from '@docusaurus/Translate';
 import type {Props} from '@theme/DocTagDocListPage';
