+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DocsVersionDropdownNavbarItem.tsx
@@ -16,7 +16,7 @@ import {
 import type {Props} from '@theme/NavbarItem/DocsVersionDropdownNavbarItem';
 import {useDocsPreferredVersion} from '@docusaurus/theme-common';
 import {translate} from '@docusaurus/Translate';
-import type {GlobalDataVersion} from '@docusaurus/plugin-content-docs-types';
+import type {GlobalDataVersion} from '@docusaurus/plugin-content-docs';
 
 const getVersionMainDoc = (version: GlobalDataVersion) =>
   version.docs.find((doc) => doc.id === version.mainDocId)!;
