+++ packages/docusaurus-utils-common/src/applyTrailingSlash.ts
