+++ packages/docusaurus-utils/src/deps.d.ts
@@ -5,7 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-// Dependencies with missing typedefs
+declare module 'resolve-pathname' {
+  export default function resolvePathname(to: string, from?: string): string;
+}
 
 declare module '@mdx-js/runtime';
 declare module 'remark-mdx-remove-imports';
