+++ packages/docusaurus/deps.d.ts
